+++ app/app.go
@@ -1,6 +1,7 @@
 package app
 
 const (
-	Name    = "circumflex"
-	Version = "2.7-dev"
+	Name               = "circumflex"
+	Version            = "2.7-dev"
+	MinimumLessVersion = 608
 )